+++ public/theme/boost_union/CHANGES.md
@@ -4,6 +4,17 @@ moodle-theme_boost_union
 Changes
 -------
 
+### v5.1-r11
+
+* 2026-04-26 - Feature: Allow not only internal WAYF services based on the Shibboleth config, but also based on JavaScript code, resolves #1230
+* 2026-04-08 - Improvement: Tweak the edit button colors so that it better fits onto tinted navbars.
+* 2026-04-08 - Improvement: Add dedicated color picker for setting the navbar color, resolves #1224
+* 2026-04-08 - Improvement: Filter additionalhtmlfooter as does Moodle core from Moodle 5.2 on (MDL-88210 / MDL-85498), resolves #847
+* 2026-04-08 - Bugfix: The login page label settings lacked a hide_if, resolves #1229
+* 2026-04-08 - Feature: Setting to split multiple IDP login buttons into individual tabs / accordeon elements / vertical sections, resolves #1130
+* 2026-04-08 - Feature: Setting to show Shibboleth internal WAYF (IdP list) on the main login page instead of the default Shibboleth button, resolves #1130
+* 2026-04-08 - Improvement: Add admin setting to tint grey colors, resolves #1220
+
 ### v5.1-r10
 
 * 2026-04-07 - Tests: Add a test to view a user tour